move props interface back into component
